requirements.txt
boto3==1.26.23
botocore==1.29.23
numpy==1.23.5
packaging==21.3
pandas==1.5.2
streamlit==1.15.2
Pillow==9.3.0